perf: eliminate OOM for 1.5M-node scans
Three memory fixes:
- u128 → u64 for all timestamp fields (modified/created/accessed/scannedAt/elapsedMs):
  saves 24 bytes × node_count (36 MB for 1.5M nodes)
- Drop full ScanResult clone from progress callback: progress ticks now send a
  lightweight {"scanning":true,"nodeCount":N,"elapsedMs":E} line instead of
  cloning ~450 MB and building another ~450 MB JSON string every 3 seconds
- DupeFilter.date_from/date_to narrowed to u64 to match

+++ src/model.rs
@@ -30,9 +30,9 @@ pub(crate) struct NodeRecord {
     pub(crate) allocated: u64,
     pub(crate) files: u64,
     pub(crate) folders: u64,
-    pub(crate) modified_ms: u128,
-    pub(crate) created_ms: u128,
-    pub(crate) accessed_ms: u128,
+    pub(crate) modified_ms: u64,
+    pub(crate) created_ms: u64,
+    pub(crate) accessed_ms: u64,
     pub(crate) depth: usize,
     pub(crate) errors: u64,
     pub(crate) children: Vec<usize>,
@@ -48,8 +48,8 @@ pub(crate) struct ScanError {
 #[derive(Clone, Debug)]
 pub(crate) struct ScanResult {
     pub(crate) root_path: String,
-    pub(crate) scanned_at_ms: u128,
-    pub(crate) elapsed_ms: u128,
+    pub(crate) scanned_at_ms: u64,
+    pub(crate) elapsed_ms: u64,
     pub(crate) thread_count: usize,
     pub(crate) nodes: Vec<NodeRecord>,
     pub(crate) errors: Vec<ScanError>,
